django_in_a_box/api/requirements.txt
asgiref==3.4.1
Django==3.2.8
djangorestframework==3.12.4
neat-python==0.92
pytz==2021.3
psycopg2-binary==2.8.5
sqlparse==0.4.2
